[Feat] data 모듈 삭제 1. data 모듈 삭제 후, Feature 모듈 패키지로 변경
