+++ .idea/compiler.xml
@@ -8,6 +8,7 @@
       <module name="commerce.core1.main" target="11" />
       <module name="commerce.core1.test" target="11" />
       <module name="commerce.core1.ui" target="21" />
+      <module name="commerce.data.home" target="21" />
       <module name="commerce.data.main" target="11" />
       <module name="commerce.data.test" target="11" />
       <module name="commerce.domain.home" target="21" />
