--- data/home/src/main/java/com/chan/home/entity/PriceEntity.kt
@@ -1,7 +1,0 @@
-package com.chan.home.entity
-
-data class PriceEntity(
-    val discountPercent: Int?,
-    val discountedPrice: Int,
-    val originalPrice: Int,
-)
