+++ feature/home/src/main/java/com/chan/home/data/dao/HomeBannerDao.kt
@@ -1,0 +1,20 @@
+package com.chan.home.data.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.chan.home.data.entity.home.HomeBannerEntity
+
+@Dao
+interface HomeBannerDao {
+
+    @Query("SELECT * FROM homeBanner")
+    suspend fun getBannerAll(): List<HomeBannerEntity>
+
+    @Insert(onConflict = OnConflictStrategy.Companion.REPLACE)
+    suspend fun insertAll(banners: List<HomeBannerEntity>)
+
+    @Query("DELETE FROM homeBanner")
+    suspend fun deleteAll()
+}
