+++ feature/home/src/main/java/com/chan/home/data/datasource/HomePopularItemDataSource.kt
@@ -1,0 +1,29 @@
+package com.chan.home.data.datasource
+
+import android.content.Context
+import com.chan.home.data.entity.home.HomePopularItemEntity
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+import dagger.hilt.android.qualifiers.ApplicationContext
+import javax.inject.Inject
+import kotlin.io.bufferedReader
+import kotlin.io.readText
+import kotlin.io.use
+
+class HomePopularItemDataSource @Inject constructor(
+    @ApplicationContext private val context: Context
+) {
+    fun getDummyPopularItems(): List<HomePopularItemEntity> {
+        return try {
+            val jsonString = context.assets.open("home_popular_item.json")
+                .bufferedReader()
+                .use { it.readText() }
+
+            val type = object : TypeToken<List<HomePopularItemEntity>>() {}.type
+            Gson().fromJson(jsonString, type)
+        } catch (e: Exception) {
+            e.printStackTrace()
+            emptyList()
+        }
+    }
+}
