+++ feature/home/src/main/java/com/chan/home/data/entity/PriceEntity.kt
@@ -1,0 +1,7 @@
+package com.chan.home.data.entity
+
+data class PriceEntity(
+    val discountPercent: Int?,
+    val discountedPrice: Int,
+    val originalPrice: Int,
+)
