+++ settings.gradle.kts
@@ -24,8 +24,6 @@ dependencyResolutionManagement {
 rootProject.name = "commerce"
 include(":app")
 
-include(":data:home")
-
 include(":feature")
 include(":feature:home")
 include(":feature:category")
